feat(customer): allow getting customer by customerId
This commit adds support for getting a customer by customerId in addition to
the existing support for getting a customer by username. The `CustomerServiceGet`
function now accepts an object with either a `username` or `customerId` property
to filter the customer to be returned.

The `CustomerControllerGet` function has been updated to use the new `identifier`
query parameter instead of the old `username` parameter. The `CustomerServiceGetSchema`
has also been updated to reflect this change.

The OpenAPI specification for the `GET /customer` endpoint has been updated to
reflect that the `identifier` parameter can now be either a `username` or a `customerId`.