feat(openapi): add PUT endpoint for updating customer information
This commit adds a new PUT endpoint to update customer information. The
endpoint is located at `/customer/{identifier}/update` and allows for
updating all fields of a customer. The endpoint is secured and requires
authentication.

The commit also adds a new `images` field to the customer schema, which
allows for storing a profile image URL. The `socialUrls` field has been
renamed to `social` to better reflect its contents.

Finally, the commit updates the `customerUpsert` endpoint to restrict
the fields that can be updated.