refactor(customer): rename socialUrls to social and avatar to images.profile.url
This commit renames the `socialUrls` property in the customer schema to `social` to better reflect the content of the object. It also renames the `avatar` property to `images.profile.url` to better reflect the structure of the object.

feat(customer): add support for images in different sizes

This commit adds support for images in different sizes by adding `width` and `height` properties to the `images.profile` object in the customer schema.

feat(customer): add support for active property

This commit adds support for an `active` property in the customer schema to indicate whether a customer is active or not.

refactor(customer): remove pattern validation for username

This commit removes the pattern validation for the `username` property in the customer schema.

refactor(customer): change images property to object

This commit changes the `images` property in the customer schema to an object with a