refactor(location): move location services to a separate file and export them from there
feat(location.service): add LocationServiceCreate, LocationServiceUpdate, LocationServiceGetCoordinates, LocationServiceGetTravelTime, and LocationServiceValidateAddress functions

This commit adds several functions to the location service. The LocationServiceCreate function creates a new location document in the database. The LocationServiceUpdate function updates an existing location document in the database. The LocationServiceGetCoordinates function retrieves the longitude and latitude of a given address. The LocationServiceGetTravelTime function retrieves the travel time and distance between two addresses. The LocationServiceValidateAddress function validates a given address.