feat(customer): add get all and get one location endpoints
Add two new endpoints to the customer function to get all locations and get one location. The `customerLocationList` endpoint is a GET request that returns all locations for a customer. The `customerLocationEdit` endpoint is a PUT request that returns a single location for a customer. The `CustomerLocationControllerGetAll` and `CustomerLocationControllerGetOne` functions handle the requests respectively. The `UserServiceGetLocations` function is added to the user service to retrieve all locations for a customer. The `CustomerLocationServiceGetOne` function is added to the customer location service to retrieve a single location for a customer.