feat(location): add locationType field to LocationServiceCreate function
Add locationType field to LocationServiceCreate function in create-destination.ts and create-origin.ts files. The locationType field is set to either LocationTypes.DESTINATION or LocationTypes.ORIGIN depending on the file.

Also, update the LocationOriginZodSchema and LocationDestinationZodSchema to extend from the LocationZodSchema. Update the LocationOrigin and LocationDestination types to include the LocationZodSchema and the respective schema. Finally, add the NumberOrStringType to the LocationDestinationZodSchema to allow for either a number or string to be used for the minDistanceForFree, distanceHourlyRate, and fixedRatePerKm fields.